requirements/production.txt
# Pro-tip: Try not to put anything here. There should be no dependency in
#	production that isn't in development.
-r base.txt



# WSGI Handler
# ------------------------------------------------

# there's no python 3 support in stable, have to use the latest release candidate for gevent
gevent==1.1rc1

gunicorn==19.4.1

# Static and Media Storage
# ------------------------------------------------
boto==2.38.0
django-storages-redux==1.3
Collectfast==0.2.3

# Mailgun Support
# ---------------
django-mailgun==0.8.0






